Grab sequence from dump file
- The sequence files produced by the pipeline were wrong in rare cases.
For species with supercontigs on the reverse strand, the logic to
calculate the reverse complement was wrong and only reverse complemented
the contigs, but not the supercontig.
Instead of creating the sequence files, we now use the sequence files
that are available as part of the "MySQL dump" pipeline.
This simplifies the pipeline significantly.